Remove MUSL CI job - incompatible with glibc-based GStreamer
MUSL cross-compilation cannot work because GStreamer system packages
are built for glibc, and the glibc headers use types (like __time64_t)
that don't exist in MUSL's libc.

True MUSL support would require either MUSL-compatible GStreamer packages
or restructuring the package to support header-only compilation.

+++ Sources/GStreamer/Pipeline.swift
@@ -597,7 +597,7 @@ public final class Pipeline: @unchecked Sendable {
     /// ```
     @discardableResult
     public func step(amount: UInt64, format: StepFormat = .buffers, rate: Double = 1.0, flush: Bool = true) -> Bool {
-        let event = gst_event_new_step(format.gstFormat, amount, rate, flush ? 1 : 0, 0)
+        let event = gst_event_new_step(format.gstFormat, guint64(amount), rate, flush ? 1 : 0, 0)
         return gst_element_send_event(_element, event) != 0
     }
